Split into multiple files
(Has bugs, will move to new development branch)
